% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.locate_closest_set}
\alias{.locate_closest_set}
\title{Locate the closest set of items to a given query}
\usage{
.locate_closest_set(q, seq_num, explored_df, set_n = NULL)
}
\arguments{
\item{q}{a named vector of query values}

\item{seq_num}{a sequence of integers indicating the possible set sizes}

\item{explored_df}{a data frame containing previously explored sets}

\item{set_n}{an optional integer indicating the number of sets to return}
}
\value{
a data frame of the closest set(s) of items to the query
}
\description{
Locate the closest set of items to a given query
}
