% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tcga.R
\name{get_tcga}
\alias{get_tcga}
\title{Extract TCGA clinical and omics data}
\usage{
get_tcga(
  tumor_type,
  gene_names = NULL,
  type = c("clinical", "expression", "CNA call", "CNA value", "mutation", "methylation",
    "protein")[1],
  short_clinical = T
)
}
\arguments{
\item{tumor_type}{A character of TCGA study abbreviation. Refer to https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations. One exception is colorectal cancer is combination of COAD and READ, labeled "COADREAD".}

\item{gene_names}{A character vector to define hugoGeneSymbol if type is not "clinical". Default is NULL}

\item{type}{A character to define type of data to extract. must be one of "clinical", "expression", "CNA call", "CNA value", "mutation", "methylation" and "protein". Default is "clinical"}

\item{short_clinical}{A Boolean vector to define whether shorten the clinical data to include just patient_id, sample_type, sample_id, os_months, os_status, age, race, sex.}
}
\value{
a data.frame for given data type
}
\description{
Extract TCGA clinical and omics data
}
\examples{
get_tcga(tumor_type="BRCA") # get clinical
get_tcga(tumor_type="BRCA",
    gene_names=c("BRCA1", "BRCA2"),
    type="expression") # get gene expression for BRCA1 and BRCA2

}
