% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HNSC-data.R
\docType{data}
\name{HNSC}
\alias{HNSC}
\title{Head & Neck Squamous Cell Carcinomas (HNSC) survival data}
\format{
a data.frame with 23 columns representing clinical data and
}
\source{
TCGA-HNSC https://www.cbioportal.org/study/summary?id=hnsc_tcga_pan_can_atlas_2018
}
\usage{
data(HNSC)
}
\description{
Data is derived from the primary tumor in HNSC from TCGA. The gene expression of \emph{TNFRSF1B} is from RNA-seq FPKQ. The cell type value (from \code{t_nk_cell:CD4+ T} to \code{myeloid:Pro-inflammatory}) is calculated from canonical cell markers using \code{GSVA} to represent the enrichment of given cell type/state in individual tumor types.
}
\examples{
data(HNSC)
}
\references{
Hoadley KA, Yau C, Hinoue T, Wolf DM, Lazar AJ, Drill E, Shen R, Taylor AM, Cherniack AD, Thorsson V, Akbani R, Bowlby R, Wong CK, Wiznerowicz M, Sanchez-Vega F, Robertson AG, Schneider BG, Lawrence MS, Noushmehr H, Malta TM; Cancer Genome Atlas Network; Stuart JM, Benz CC, Laird PW. Cell-of-Origin Patterns Dominate the Molecular Classification of 10,000 Tumors from 33 Types of Cancer. Cell. 2018 Apr 5;173(2):291-304.e6. doi: 10.1016/j.cell.2018.03.022. PMID: 29625048; PMCID: PMC5957518.
}
\keyword{datasets}
