% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxph_1d2b_optimal.R
\name{run_coxph_1d2b_optimal}
\alias{run_coxph_1d2b_optimal}
\title{Run Cox proportional hazards model with one continuous variable using optimal two-boundary cutoffs}
\usage{
run_coxph_1d2b_optimal(
  formula,
  input_data,
  q_col,
  low_q = 0.2,
  high_q = 0.8,
  breaks = NULL,
  by = NULL,
  pvalue_type = c("log", "sc", "wald")[1],
  min_group_size = 10,
  baseline = NULL,
  method = c("median-anchored", "exhaustive")[1],
  set_n = 3
)
}
\arguments{
\item{formula}{A formula object specifying the survival object and
covariates to include in the regression model.}

\item{input_data}{A data frame containing the variables specified in the formula.}

\item{q_col}{A character string specifying the name of the first continuous variable in
\code{input_data} to regress on in  survival \code{formula}}

\item{low_q}{the lower quantile for generating the search space for optimal cutoff values (default: 0.2)}

\item{high_q}{the upper quantile for generating the search space for optimal cutoff values (default: 0.8)}

\item{breaks}{the number of intervals to divide the search space into (if \code{by} is not specified)}

\item{by}{the width of each interval in the search space (if \code{breaks} is not specified)}

\item{pvalue_type}{the type of p-value to be used for ranking the results (default: "log")}

\item{min_group_size}{the minimum number of observations required for each group (default: 10)}

\item{baseline}{An optional character string specifying the method for
estimating the baseline hazard function. Must be one of "low",
"high". Defaults to NULL, which uses "high".}

\item{method}{the method for generating the search space for optimal cutoff values (default: "median-anchored")}

\item{set_n}{The number of quantitle combinations closest to the previous set of cutoff generating lowest p-value for the median-anchored method. Default is 3.}
}
\value{
A list with the following elements:
\item{best_test}{The output of the coxph function for the best combination of cutoffs.}
\item{best_data}{The input data for the best combination of cutoffs.}
\item{best_q}{The best combination of cutoffs.}
\item{rank_q}{A data frame with the p-value of the score test for each combination of cutoffs.}
}
\description{
Similar to \code{run_coxph_1d_optimal}, the funciton performs coxph survival analysis by converting single continuous variable to categorical variable with optimal quantile cutoff. However, the optimal cutoff have two boundaries now, including upper and lower quantile cutoff. "median-anchored method" controls upper quantile cutoff and starts with 0.5.
}
\examples{
data(HNSC)
run_coxph_1d2b_optimal(
    formula=Surv(SURVIVAL_os_days, SURVIVAL_os_event) ~ TNFRSF1B,
    input_data = HNSC,
    q_col="TNFRSF1B",
    by = 0.1)


}
