% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxph_1d_optimal.R
\name{run_coxph_1d_optimal}
\alias{run_coxph_1d_optimal}
\title{Find optimal quantile cutoff for Cox proportional hazards regression}
\usage{
run_coxph_1d_optimal(
  formula,
  input_data,
  q_col,
  low_q = 0.2,
  high_q = 0.8,
  breaks = NULL,
  by = NULL,
  pvalue_type = c("log", "sc", "wald")[1],
  min_group_size = 10,
  baseline = NULL
)
}
\arguments{
\item{formula}{A formula object specifying the survival object and
covariates to include in the regression model.}

\item{input_data}{A data frame containing the variables specified in the formula.}

\item{q_col}{A character string specifying the name of the first continuous variable in
\code{input_data} to regress on in  survival \code{formula}}

\item{low_q}{A numeric value specifying the lowest quantile cutoff. Defaults to 0.2.}

\item{high_q}{A numeric value specifying the highest quantile cutoff. Defaults to 0.8.}

\item{breaks}{An integer specifying the number of quantile cutoffs to use
if \code{by} is not specified. Ignored if \code{by} is specified. Defaults to NULL.}

\item{by}{A numeric value specifying the increment between quantile cutoffs
to use if \code{breaks} is not specified. Ignored if \code{breaks} is specified.
Defaults to NULL.}

\item{pvalue_type}{A character string specifying the type of p-value to use
when ranking the quantile cutoffs. Must be one of "log", "sc", or "wald".
Defaults to "log".}

\item{min_group_size}{An integer specifying the minimum number of
observations in each group for the regression to be performed. Defaults
to 10.}

\item{baseline}{An optional character string specifying the method for
estimating the baseline hazard function. Must be one of "low",
"high". Defaults to NULL, which uses "high".}
}
\value{
A list with the following components:
\describe{
\item{best_test}{The coxph test object with the lowest p-value.}
\item{best_data}{The data frame used to fit the best test.}
\item{best_q}{The quantile cutoff value that produced the best test.}
\item{rank_q}{A data frame with the ranked quantile cutoff values and
their corresponding p-values for the selected p-value type.}
}
}
\description{
This function performs Cox proportional hazards regression with multiple
quantile cutoff values and returns the best test result, best data, best
quantile cutoff, and a dataframe of the quantile cutoffs ranked by the
selected p-value type.
}
\examples{
data(HNSC)
run_coxph_1d_optimal(
    formula=Surv(SURVIVAL_os_days, SURVIVAL_os_event) ~ TNFRSF1B,
    input_data = HNSC,
    q_col="TNFRSF1B",
    by = 0.1)
}
