% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxph_2d.R
\name{run_coxph_2d}
\alias{run_coxph_2d}
\title{Run Cox proportional hazards regression with two continuous variables}
\usage{
run_coxph_2d(
  formula,
  input_data,
  q1_col,
  q2_col,
  q1 = 0.5,
  q1_2 = NULL,
  q2 = 0.5,
  q2_2 = NULL,
  min_group_size = 10,
  combine_group = NULL,
  baseline = NULL
)
}
\arguments{
\item{formula}{A formula specifying the model to be fitted.}

\item{input_data}{A data.frame containing the variables specified in the formula.}

\item{q1_col}{A character string specifying the name of the first continuous variable.}

\item{q2_col}{A character string specifying the name of the second continuous variable.}

\item{q1}{A numeric value specifying the quantile cutoff for the first continuous variable.}

\item{q1_2}{A numeric value specifying the upper quantile cutoff for the first continuous variable.
Set to NULL or NA if no upper cutoff is desired.}

\item{q2}{A numeric value specifying the quantile cutoff for the second continuous variable.}

\item{q2_2}{A numeric value specifying the upper quantile cutoff for the second continuous variable.
Set to NULL or NA if no upper cutoff is desired.}

\item{min_group_size}{An integer specifying the minimum number of observations in each group.}

\item{combine_group}{A character vector specifying the group levels to be combined into "other". It must be from "high:high", "high:low", "low:high" and "low:low".}

\item{baseline}{A character string specifying the group level to be used as the baseline for the analysis.  It must be "high:high", "high:low", "low:high", "low:low" or 'other'.}
}
\value{
A list with two elements: "test", which contains the results of the Cox proportional hazards regression,
and "data", which contains the data used for the analysis.
}
\description{
Perform coxph survival analysis by converting two continuous variables to categorical variables with user defined quantile cutoff.
}
\examples{
data(HNSC)
run_coxph_2d(
   formula=Surv(SURVIVAL_os_days, SURVIVAL_os_event) ~
       TNFRSF1B + `t_nk_cell:Treg` + strata(cancer_grading),
   input_data = HNSC,
   q1_col="TNFRSF1B",
  q2_col = "t_nk_cell:Treg",
  q1=0.4, q1_2 = 0.6, q2 = 0.5)

}
