% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxph_2d2b_optimal.R
\name{run_coxph_2d2b_optimal}
\alias{run_coxph_2d2b_optimal}
\title{Run Cox proportional hazards model with two continuous variables using optimal two-boundary cutoffs}
\usage{
run_coxph_2d2b_optimal(
  formula,
  input_data,
  q1_col,
  q2_col,
  low_q = 0.2,
  high_q = 0.8,
  breaks = NULL,
  by = NULL,
  pvalue_type = c("log", "sc", "wald")[1],
  min_group_size = 10,
  combine_group = NULL,
  baseline = NULL,
  method = c("median-anchored", "exhaustive")[1],
  set_n = 3
)
}
\arguments{
\item{formula}{a formula specifying the survival object and covariates of interest}

\item{input_data}{a data.frame containing the survival object and covariates specified in \code{formula}}

\item{q1_col}{the name of the column in \code{input_data} containing the first continuous covariate}

\item{q2_col}{the name of the column in \code{input_data} containing the second continuous covariate}

\item{low_q}{the lower quantile for generating the search space for optimal cutoff values (default: 0.2)}

\item{high_q}{the upper quantile for generating the search space for optimal cutoff values (default: 0.8)}

\item{breaks}{the number of intervals to divide the search space into (if \code{by} is not specified)}

\item{by}{the width of each interval in the search space (if \code{breaks} is not specified)}

\item{pvalue_type}{the type of p-value to be used for ranking the results (default: "log")}

\item{min_group_size}{the minimum number of observations required for each group (default: 10)}

\item{combine_group}{A character vector specifying the group levels to be combined into "other". It must be from "high:high", "high:low", "low:high" and "low:low". Default is NULL, meaning no combination.}

\item{baseline}{A character string specifying the group level to be used as the baseline for the analysis.  It must be "high:high", "high:low", "low:high", "low:low" or 'other'. Default is NULL, meaning using "high:high" as baseline.}

\item{method}{the method for generating the search space for optimal cutoff values (default: "median-anchored")}

\item{set_n}{The number of quantitle combinations closest to the previous set of cutoff generating lowest p-value for the median-anchored method. Default is 3.}
}
\value{
A list with the following elements:
\item{best_test}{The output of the coxph function for the best combination of cutoffs.}
\item{best_data}{The input data for the best combination of cutoffs.}
\item{best_q}{The best combination of cutoffs.}
\item{rank_q}{A data frame with the p-value of the score test for each combination of cutoffs.}
}
\description{
Similar to \code{run_coxph_2d_optimal}, the funciton performs coxph survival analysis by converting two continuous variables to categorical variables with optimal quantile cutoffs. However, the optimal cutoffs have two boundaries now, including upper and lower quantile cutoff. "median-anchored method" controls second variable and starts with single median cutoff.
}
\examples{
data(HNSC)
run_coxph_2d2b_optimal(
    formula=Surv(SURVIVAL_os_days, SURVIVAL_os_event) ~
        TNFRSF1B + `t_nk_cell:Treg` + strata(cancer_grading),
    input_data = HNSC,
    q1_col="TNFRSF1B",
    q2_col = "t_nk_cell:Treg",
    by=0.1)

run_coxph_2d2b_optimal(
    formula=Surv(SURVIVAL_os_days, SURVIVAL_os_event) ~
        TNFRSF1B + `t_nk_cell:Treg` + strata(cancer_grading),
    input_data = HNSC,
    q1_col="TNFRSF1B",
    q2_col = "t_nk_cell:Treg",
    by=0.2,
    combine_group = c("low:high", "high:low"),
    method = "exhaustive")

}
