% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxph_2d_optimal.R
\name{run_coxph_2d_optimal}
\alias{run_coxph_2d_optimal}
\title{Run Cox proportional hazards model with two continuous variables using optimal cutoffs}
\usage{
run_coxph_2d_optimal(
  formula,
  input_data,
  q1_col,
  q2_col,
  low_q = 0.2,
  high_q = 0.8,
  breaks = NULL,
  by = NULL,
  pvalue_type = c("log", "sc", "wald")[1],
  min_group_size = 10,
  combine_group = NULL,
  baseline = NULL,
  method = c("median-anchored", "exhaustive")[1],
  set_n = 3
)
}
\arguments{
\item{formula}{A formula specifying the model to be fitted.}

\item{input_data}{A data frame containing the variables specified in the formula.}

\item{q1_col}{A character string specifying the name of the first continuous variable.}

\item{q2_col}{A character string specifying the name of the second continuous variable.}

\item{low_q}{The lower limit of the quantile range used to generate the search space for optimal cutoffs.}

\item{high_q}{The upper limit of the quantile range used to generate the search space for optimal cutoffs.}

\item{breaks}{The number of quantiles used to generate the search space for optimal cutoffs. Only used if by is NULL.}

\item{by}{The distance between the quantiles used to generate the search space for optimal cutoffs. Only used if breaks is NULL.}

\item{pvalue_type}{The type of p-value used for ranking optimal cutoffs. Can be "log", "sc", or "wald".}

\item{min_group_size}{An integer specifying the minimum number of observations in each group. Default is 10}

\item{combine_group}{A character vector specifying the group levels to be combined into "other". It must be from "high:high", "high:low", "low:high" and "low:low". Default is NULL, meaning no combination.}

\item{baseline}{A character string specifying the group level to be used as the baseline for the analysis.  It must be "high:high", "high:low", "low:high", "low:low" or 'other'. Default is NULL, meaning using "high:high" as baseline.}

\item{method}{The method used to determine the optimal cutoffs. Can be "median-anchored" or "exhaustive".}

\item{set_n}{The number of quantitle combinations closest to the previous set of cutoff generating lowest p-value for the median-anchored method. Default is 3.}
}
\value{
A list with the following elements:
\item{best_test}{The output of the coxph function for the best combination of cutoffs.}
\item{best_data}{The input data for the best combination of cutoffs.}
\item{best_q}{The best combination of cutoffs.}
\item{rank_q}{A data frame with the p-value of the score test for each combination of cutoffs.}
}
\description{
Coxph survival analysis by converting two continuous variables to categorical variables with optimal quantile cutoffs. The optimal cutoffs are determined by exhaustively searching or using median-anchored method. The optimal cutoffs are determined by minimizing the p-value of the score test of the Cox proportional hazards model.
}
\examples{
data(HNSC)
run_coxph_2d_optimal(
    formula=Surv(SURVIVAL_os_days, SURVIVAL_os_event) ~
        TNFRSF1B + `t_nk_cell:Treg` + strata(cancer_grading),
    input_data = HNSC,
    q1_col="TNFRSF1B",
    q2_col = "t_nk_cell:Treg",
    breaks=10)

run_coxph_2d_optimal(
    formula=Surv(SURVIVAL_os_days, SURVIVAL_os_event) ~
        TNFRSF1B + `t_nk_cell:Treg` + strata(cancer_grading),
    input_data = HNSC,
    q1_col="TNFRSF1B",
    q2_col = "t_nk_cell:Treg",
    by=0.1,
    combine_group = c("low:high", "high:low"),
    baseline="other",
    set_n = 5)
}
